refactor: team folder
